backend/requirements.txt
# ── Core Framework ──────────────────────────────────────────────
flask>=3.0,<4.0
flask-cors>=4.0,<5.0
flask-limiter>=3.5,<4.0
gunicorn>=22.0,<23.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0

# ── Data & NLP ──────────────────────────────────────────────────
pandas>=2.0,<3.0
wikipedia>=1.4,<2.0
numpy>=1.24,<2.0
duckduckgo-search>=4.0

# ── LLM & Transformers ─────────────────────────────────────────
transformers==4.46.1
sentencepiece>=0.2,<1.0
sacremoses>=0.1,<1.0
torch==2.3.0
accelerate>=0.30,<1.0

# ── ML Model ───────────────────────────────────────────────────
xgboost>=2.0,<3.0
scikit-learn>=1.3,<2.0
matplotlib>=3.7,<4.0
seaborn>=0.13,<1.0

# ── RAG + LSTM ──────────────────────────────────────────────────
faiss-cpu>=1.7,<2.0
sentence-transformers>=2.2,<3.0

# ── MLOps ───────────────────────────────────────────────────────
mlflow>=2.13,<3.0

# ── Testing ─────────────────────────────────────────────────────
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
